Add Height property to BodyMeasurement
Previously, the body-fat percentage was being calculated with the values
in the BodyMeasurement and the Height property in AppUser. Now, the
body-fat percentage is calculated using the height property within the
BodyMeasurement.

I think this is a better choice because what if the app user's height
changes, or what if he/she accidentally put in the wrong height when
signing up? In these cases, there would be no way to adjust the
calculations for the previous measurements. This implementation fixes
those issues because the height is set, and therefore editable, for each
measurement.