Add tests for toast messages after a measurement has been created/edited
The added tests are commented out. For whatever reason, I could not get
these test to work when the other tests in the BodyMeasurementsPage test
file are run.

Each of these tests will pass if it is the only test that is run.
Hopefully, in the future, I can come back to these tests and get them to
pass when the other tests run